fix(YALB-1117): rename news to posts
