+++ components/02-molecules/cards/reference-card/examples/_card--examples.twig
@@ -1,5 +1,5 @@
 {% set reference_card__date__formatted %}
-  {% set date_time__format = card_example_type == 'news' ? 'date' %}
+  {% set date_time__format = card_example_type == 'post' ? 'date' %}
   {% include "@atoms/date-time/yds-date-time.twig" with {
     date_time__start: reference_card__date,
   } %}
@@ -12,10 +12,10 @@
   } %}
 {% endset %}
 
-{# News Card #}
-{% if card_example_type == 'news' %}
+{# Post Card #}
+{% if card_example_type == 'post' %}
   {% set reference_card__overline = reference_card__date__formatted %}
-  {# If news, list, and not freatured, hide snippet. #}
+  {# If post, list, and not freatured, hide snippet. #}
   {% if card_collection__type == 'list' and card_collection__featured == 'false' %}
     {% set reference_card__snippet = null %}
   {% endif %}
