+++ components/05-page-examples/post/post-article.twig
@@ -6,7 +6,7 @@
     {% include "@organisms/menu/breadcrumbs/yds-breadcrumbs.twig" with {
       breadcrumbs__items: [
         {title: 'Home', url: '#'},
-        {title: 'Chem News', url: '#'},
+        {title: 'Chem Post', url: '#'},
         {title: page_title__heading, url: '#', is_active: true},
       ],
     } %}
