+++ components/05-page-examples/post/post-grid.twig
@@ -1,6 +1,6 @@
 <div class="card-collection" data-component-width="max" data-collection-type="grid" data-collection-featured="true">
 	<div class="card-collection__inner">
-		<h2 class="card-collection__heading">News Card Grid Heading - Static HTML grid of cards</h2>
+		<h2 class="card-collection__heading">Post Card Grid Heading - Static HTML grid of cards</h2>
 		<ul class="card-collection__cards">
 			<li class="reference-card" style="cursor: pointer;">
 				<div class="reference-card__content">
@@ -32,7 +32,7 @@
 					</h3>
 
 					<div class="reference-card__snippet">
-            Et malesuada fames ac turpis egestas. Velit ut tortor pretium viverra suspendisse potenti nullam. Quis ipsum suspendisse ultrices gravida dictum fusce ut placerat orci. 
+            Et malesuada fames ac turpis egestas. Velit ut tortor pretium viverra suspendisse potenti nullam. Quis ipsum suspendisse ultrices gravida dictum fusce ut placerat orci.
           </div>
 				</div>
 				<div class="reference-card__image">
@@ -50,7 +50,7 @@
 						</a>
 					</h3>
 					<div class="reference-card__snippet">
-            Purus in mollis nunc sed id semper. Neque laoreet suspendisse interdum consectetur libero. Nibh venenatis cras sed felis. Volutpat diam ut venenatis tellus in metus vulputate eu scelerisque. Neque viverra justo nec ultrices dui sapien eget mi. Tempor orci dapibus ultrices in. Volutpat odio facilisis mauris sit amet massa vitae tortor. 
+            Purus in mollis nunc sed id semper. Neque laoreet suspendisse interdum consectetur libero. Nibh venenatis cras sed felis. Volutpat diam ut venenatis tellus in metus vulputate eu scelerisque. Neque viverra justo nec ultrices dui sapien eget mi. Tempor orci dapibus ultrices in. Volutpat odio facilisis mauris sit amet massa vitae tortor.
           </div>
 				</div>
 				<div class="reference-card__image">
