+++ components/05-page-examples/standard-pages/standard-page-with-banner-left-align.twig
@@ -11,7 +11,7 @@
         grand_hero__snippet: banner__snippet,
         grand_hero__link__content: banner__link__content,
         grand_hero__link__url: banner__link__url,
-        
+
       } %}
 	{% endif %}
 	{% include "@organisms/menu/breadcrumbs/yds-breadcrumbs.twig" %}
@@ -35,7 +35,7 @@
   } %}
 
 	{% include "@molecules/text/yds-text-field.twig" with {
-      text_field__alignment: 'left', 
+      text_field__alignment: 'left',
       text_field__content: '<p>Chemistry has been responsible for some of the most significant improvements in our quality of life over the last century.</p><p>The discovery of antibiotics and other pharmaceuticals, the advent of computers, and the development of industrial methods to produce fertilizer, all have required fundamental advances in chemistry. Chemistry is increasingly playing a central role in the development of alternative-energy vectors to replace fossil fuels, the realization of practical quantum computers, the discovery of new methods to treat and prevent diseases, and the adoption of more sustainable industrial processes.</p>',
     } %}
 
@@ -46,7 +46,7 @@
     } %}
 
 	{% include "@molecules/text/yds-text-field.twig" with {
-      text_field__alignment: 'left', 
+      text_field__alignment: 'left',
       text_field__content: '<p>Chemistry has been responsible for some of the most significant improvements in our quality of life over the last century.</p><p>The discovery of antibiotics and other pharmaceuticals, the advent of computers, and the development of industrial methods to produce fertilizer, all have required fundamental advances in chemistry. Chemistry is increasingly playing a central role in the development of alternative-energy vectors to replace fossil fuels, the realization of practical quantum computers, the discovery of new methods to treat and prevent diseases, and the adoption of more sustainable industrial processes.</p>',
     } %}
 
@@ -60,7 +60,7 @@
 	{% endembed %}
 
 	{% include "@molecules/text/yds-text-field.twig" with {
-      text_field__alignment: 'left', 
+      text_field__alignment: 'left',
       text_field__content: '<p>Chemistry has been responsible for some of the most significant improvements in our quality of life over the last century.</p><p>The discovery of antibiotics and other pharmaceuticals, the advent of computers, and the development of industrial methods to produce fertilizer, all have required fundamental advances in chemistry. Chemistry is increasingly playing a central role in the development of alternative-energy vectors to replace fossil fuels, the realization of practical quantum computers, the discovery of new methods to treat and prevent diseases, and the adoption of more sustainable industrial processes.</p>',
     } %}
 
@@ -72,16 +72,16 @@
     } %}
 
 	{% include "@molecules/text/yds-text-field.twig" with {
-      text_field__alignment: 'left', 
+      text_field__alignment: 'left',
       text_field__content: '<p>Chemistry has been responsible for some of the most significant improvements in our quality of life over the last century.</p><p>The discovery of antibiotics and other pharmaceuticals, the advent of computers, and the development of industrial methods to produce fertilizer, all have required fundamental advances in chemistry. Chemistry is increasingly playing a central role in the development of alternative-energy vectors to replace fossil fuels, the realization of practical quantum computers, the discovery of new methods to treat and prevent diseases, and the adoption of more sustainable industrial processes.</p>',
     } %}
 
 	{% include "@molecules/image/yds-content-image.twig" with {
       content_image__width:'site',
-      content_image__alignment: 'left', 
+      content_image__alignment: 'left',
     }%}
 	{% include "@molecules/text/yds-text-field.twig" with {
-      text_field__alignment: 'left', 
+      text_field__alignment: 'left',
       text_field__content: '<p>The undergraduate program in Chemistry at Yale reflects the position of chemistry as the foundation of all the molecular sciences. Students are equipped with the technical skills to appreciate the scientific basis for previous discoveries and develop the fundamental expertise required to make future breakthroughs. Under the tutelage of world-leading researchers, students are exposed to a broad range of topics in chemistry. The development of technical skills through lecture classes is complemented with hands-on experience in state-of-the-art chemistry laboratories.</p><p>Many students also perform independent laboratory research under the guidance of a faculty mentor. This rigorous training prepares students for professional careers in a diverse array of fields by teaching them how to apply the scientific method, providing them with skills in quantitative reasoning, and exposing them to scientific research.</p><p>After graduation, students with a B.A. or B.S. degree often pursue work or further studies in chemistry, biochemistry, or health-related disciplines, but also find their broad scientific training beneficial in energy research, policy, environment, business management, and law. As the problems of contemporary society involve ever more complex scientific issues, degree programs in the sciences become increasingly appropriate for students wishing to pursue careers in public policy, government, or public service.</p>',
     } %}
 	{% include "@molecules/video/yds-video.twig" with {
@@ -103,7 +103,7 @@
 	{% include "@atoms/divider/yds-divider.twig" %}
 
 	{% include "@organisms/card-collection/yds-card-collection.twig" with {
-      card_example_type: 'news',
+      card_example_type: 'post',
       card_collection__cards:[
         {
           reference_card__date: '2022-03-30 13:00',
@@ -134,7 +134,7 @@
 	{% include "@molecules/accordion/yds-accordion.twig" with {
       accordion__heading: 'Accordion Group - left align',
       accordion__width: 'site',
-      accordion__alignment: 'left', 
+      accordion__alignment: 'left',
       accordion__items: [
         {
           accordion__item__heading: accordion__item__heading,
@@ -162,7 +162,7 @@
 
 	{% embed "@organisms/component-wrapper/yds-component-wrapper.twig" with {
     component_wrapper__width: 'site',
-    component_wrapper__alignment: 'left', 
+    component_wrapper__alignment: 'left',
     component_wrapper__label: 'Contact',
   } %}
 		{% block component_wrapper_inner %}
