+++ components/05-page-examples/standard-pages/standard-page-with-banner.twig
@@ -78,7 +78,7 @@
 	{% include "@atoms/divider/yds-divider.twig" %}
 
 	{% include "@organisms/card-collection/yds-card-collection.twig" with {
-      card_example_type: 'news',
+      card_example_type: 'post',
       card_collection__cards:[
         {
           reference_card__date: '2022-03-30 13:00',
